requirements.txt
streamlit>=1.20.0
openai>=0.27.0
python-dotenv>=1.0.0
pyttsx3>=2.90
fpdf2>=2.6.0
plotly>=5.0.0
numpy>=1.24.0
pandas>=1.5.0
PyPDF2>=3.0.0
